basic file transfer functionality is complete

+++ src/edu/isi/usaid/pifi/Constants.java
@@ -54,8 +54,8 @@ public class Constants {
 	public static final short META_FROM_MASTER = 1003;
 	
 	
-	public static final int OK_RESPONSE = 200;
-	public static final int FAIL_RESPONSE = 400;
+	public static final short OK_RESPONSE = 200;
+	public static final short FAIL_RESPONSE = 400;
 	
 
 }
